feat: add platform-version CLI argument
ECS is still deploying to Fargate 1.3.0 by default. This is
pretty old, and is missing some very useful things.

This PR adds the ability to specify the platform version via
the new `--platform-version` argument. If specified, and if
the `--fargate` option has also been specified, the provided
platform version will be used to launch the task.

E.g.:

ecs-run-task --fargate --platform-version 1.4.0 ...